Learn about mongodb relationship and how to connect two tables for foreign key

+++ JournalApp/src/main/java/com/kathan/JournalApp/controller/JournalControllerDB.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,7 +17,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.kathan.JournalApp.entities.JournalEntry;
+import com.kathan.JournalApp.entities.Users;
 import com.kathan.JournalApp.service.JournalService;
+import com.kathan.JournalApp.service.UserService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -25,15 +28,26 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class JournalControllerDB {
 	private final JournalService service;
+	private final UserService userService;
 
 	
-	@PostMapping
-	public ResponseEntity<JournalEntry> create(@RequestBody JournalEntry j) {
+	@PostMapping("{username}")
+	@Transactional
+	public ResponseEntity<JournalEntry> createJournalEntryForUser(@RequestBody JournalEntry j,@PathVariable String username) {
 		try {
 			if(j.getDate()==null) {
 				j.setDate(LocalDate.now());
 			}
-			service.saveEntry(j);
+			service.saveEntry(j); //Saving entry in Journal Entry database
+			
+			Users user = userService.findByUsername(username);
+			if(user==null) {
+				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			}
+			
+			user.getJournal_entries().add(j); //Saving entry with user
+			userService.saveEntry(user); //Now saving user into database
+			
 			return new ResponseEntity<>(j,HttpStatus.CREATED);
 		}catch(Exception e) {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -41,10 +55,15 @@ public class JournalControllerDB {
 	
 	}
 	
-	@GetMapping
-	public ResponseEntity<List<JournalEntry>> getAll(){
-		//return service.getAll();	
-		List<JournalEntry> allEntry = service.getAll();
+	@GetMapping("{username}")
+	public ResponseEntity<List<JournalEntry>> getAllJournalEntriesOfUser(@PathVariable String username){
+		Users user = userService.findByUsername(username);
+		if(user==null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		
+		List<JournalEntry> allEntry = user.getJournal_entries();
+		
 		if(allEntry==null || allEntry.isEmpty()) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
@@ -60,10 +79,12 @@ public class JournalControllerDB {
 	
 	@PutMapping("/id/{myId}")
 	public ResponseEntity<JournalEntry> updateById(@PathVariable String myId,@RequestBody JournalEntry newEntry) {
+
 		Optional<JournalEntry> oldEntry = service.getById(myId);
 		if(oldEntry.isEmpty()) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
+		
 		JournalEntry old = oldEntry.get();
 		old.setTitle(newEntry.getTitle()==null ? old.getTitle() : newEntry.getTitle());
 		old.setContent(newEntry.getContent()==null ? old.getContent() : newEntry.getContent());
@@ -72,8 +93,16 @@ public class JournalControllerDB {
 		return new ResponseEntity<>(old,HttpStatus.OK);
 	}
 	
-	@DeleteMapping("/id/{myId}")
-	public ResponseEntity<?> deleteById(@PathVariable String myId) {
+	@DeleteMapping("/{myId}/{username}")
+	@Transactional
+	public ResponseEntity<?> deleteById(@PathVariable String myId,@PathVariable String username) {
+		Users user = userService.findByUsername(username);
+		if(user==null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		
+		user.getJournal_entries().removeIf(entry -> entry.getId().equals(myId));
+		userService.saveEntry(user);
 		service.deleteById(myId);
 		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
